# FSA.Tickets-Please/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SIMS incident/signals management API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Addresses/{addressId}": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "Fetch address",
        "operationId": "GetAddress",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "addressId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SimsAddressViewModel"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
    "/api/Products/{incidentSignal}": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create new Product",
        "description": "Create new Product",
        "operationId": "AddProduct",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SimsProduct"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update Product",
        "description": "Update Product",
        "operationId": "UpdateProduct",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SimsProduct"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/Products/{incidentSignal}/{productId}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Fetch product",
        "description": "Fetch product",
        "operationId": "GetProduct",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SimsProductDisplayModel"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Products/all/{incidentSignal}/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Fetch products",
        "description": "Fetch products",
        "operationId": "GetAllProducts",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SimsProduct"
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Products/Addresses/{incidentSignal}/{productId}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Fetch product addresses",
        "description": "Fetch product addresses",
        "operationId": "GetProductAddresses",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductFboAddressViewModel"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Products/Fbo/{incidentSignal}": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Assign fbo to product",
        "description": "Assign fbo to product",
        "operationId": "AddFbo",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductFboInfoViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Remove fbo from product",
        "description": "Remove fbo from product",
        "operationId": "RemoveFbo",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductFboInfoViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update fbo to product",
        "description": "Update fbo to product",
        "operationId": "UpdateFbo",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductFboInfoViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Products/Dashboard/{incidentSignal}/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Product dashboard",
        "description": "Product dashboard",
        "operationId": "GetProductDashboard",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentSignal",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 10
          },
          {
            "in": "query",
            "name": "pageNo",
            "type": "integer",
            "format": "int32",
            "default": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SimsProductDashboard"
              }
            }
          },
          "500": {
            "description": "Server Error"
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "SimsAddressViewModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "mainContact": {
          "type": "string"
        },
        "addressLine1": {
          "type": "string"
        },
        "addressLine2": {
          "type": "string"
        },
        "townCity": {
          "type": "string"
        },
        "county": {
          "type": "string"
        },
        "countryId": {
          "format": "int32",
          "type": "integer"
        },
        "postCode": {
          "type": "string"
        },
        "telephoneNumber": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "contactMethodId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": {
        "type": "object"
      }
    },
...
    "ProductPackSizeModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "size": {
          "type": "string"
        },
        "unitId": {
          "format": "int32",
          "type": "integer"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "ProductDateModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "dateTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "ProductViewModel": {
      "type": "object",
      "properties": {
        "hostId": {
          "format": "int32",
          "type": "integer"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "brand": {
          "type": "string"
        },
        "productTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "countryOfOriginId": {
          "format": "int32",
          "type": "integer"
        },
        "amount": {
          "type": "string"
        },
        "amountUnitTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "packSizes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductPackSizeModel"
          }
        },
        "productDates": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductDateModel"
          }
        },
        "batchCodes": {
          "type": "string"
        },
        "packDescription": {
          "type": "string"
        },
        "additionalInfo": {
          "type": "string"
        }
      }
    },
    "SimsProductPackSize": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "size": {
          "type": "string"
        },
        "unitId": {
          "format": "int32",
          "type": "integer"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "SimsProductDate": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "dateTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "SimsProduct": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "hostId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "brand": {
          "type": "string"
        },
        "productTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "countryOfOriginId": {
          "format": "int32",
          "type": "integer"
        },
        "amount": {
          "type": "string"
        },
        "amountUnitTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "packSizes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SimsProductPackSize"
          }
        },
        "productDates": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SimsProductDate"
          }
        },
        "batchCodes": {
          "type": "string"
        },
        "packDescription": {
          "type": "string"
        },
        "additionalInfo": {
          "type": "string"
        },
        "added": {
          "format": "date-time",
          "type": "string"
        },
        "lastUpdated": {
          "format": "date-time",
          "type": "string"
        },
        "lastUpdatedBy": {
          "type": "string"
        }
      }
    },
    "SimsProductDisplayModel": {
      "type": "object",
      "properties": {
        "dataSourceId": {
          "format": "int32",
          "type": "integer"
        },
        "signalUrl": {
          "type": "string"
        },
        "signalDataSource": {
          "type": "string"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "hostId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "brand": {
          "type": "string"
        },
        "productTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "countryOfOriginId": {
          "format": "int32",
          "type": "integer"
        },
        "amount": {
          "type": "string"
        },
        "amountUnitTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "packSizes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SimsProductPackSize"
          }
        },
        "productDates": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SimsProductDate"
          }
        },
        "batchCodes": {
          "type": "string"
        },
        "packDescription": {
          "type": "string"
        },
        "additionalInfo": {
          "type": "string"
        },
        "added": {
          "format": "date-time",
          "type": "string"
        },
        "lastUpdated": {
          "format": "date-time",
          "type": "string"
        },
        "lastUpdatedBy": {
          "type": "string"
        }
      }
    },
    "SimsFboTypes": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        4,
        8,
        16,
        32,
        64,
        128,
        256,
        512,
        1024,
        2048,
        4096,
        8192,
        16384,
        32768,
        65536
      ],
      "type": "integer"
    },
    "SimsAddressContact": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "addressId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "telephoneNumber": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "isMain": {
          "type": "boolean"
        }
      }
    },
    "ProductFboAddressViewModel": {
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "fboTypes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SimsFboTypes"
          }
        },
        "mainContact": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "addressLine1": {
          "type": "string"
        },
        "addressLine2": {
          "type": "string"
        },
        "townCity": {
          "type": "string"
        },
        "county": {
          "type": "string"
        },
        "countryId": {
          "format": "int32",
          "type": "integer"
        },
        "postCode": {
          "type": "string"
        },
        "telephoneNumber": {
          "type": "string"
        },
        "contactMethodId": {
          "format": "int32",
          "type": "integer"
        },
        "contacts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SimsAddressContact"
          }
        }
      }
    },
    "ProductFboInfoViewModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "fboTypes": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        },
        "addressId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "SimsProductDashboard": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "productType": {
          "type": "string"
        },
        "batchCodes": {
          "type": "string"
        },
        "lastUpdated": {
          "format": "date-time",
          "type": "string"
        },
        "addressNames": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}